:hammer: refactored new project template
