+++ ResourceProvisioner/test/ResourceProvisioner.Infrastructure.UnitTests/Templates/AzureStorageBlobTemplateTests.cs
@@ -295,27 +295,4 @@ public class AzureStorageBlobTemplateTests
              });
          }
      }
-
-    private static async Task<int> SetupNewProjectTemplate(string workspaceAcronym)
-    {
-        var workspace = new TerraformWorkspace
-        {
-            Acronym = workspaceAcronym
-        };
-        await _repositoryService.FetchRepositoriesAndCheckoutProjectBranch(workspaceAcronym);
-
-        var module = new TerraformTemplate
-        {
-            Name = TerraformTemplate.NewProjectTemplate,
-            Version = "latest"
-        };
-
-        await _terraformService.CopyTemplateAsync(module, workspace);
-        await _terraformService.ExtractVariables(module, workspace);
-        await _terraformService.ExtractBackendConfig(workspaceAcronym);
-
-        var moduleDestinationPath = DirectoryUtils.GetProjectPath(_configuration, workspaceAcronym);
-        return Directory
-            .GetFiles(moduleDestinationPath, "*.*", SearchOption.TopDirectoryOnly).Length;
-    }
 }
